Unify platform-specific files using generics and !linux tags
- clock: replace 3 arch-specific setters (386, 64bit, darwin) with one
  generic setTimexField[T] that handles int32/int64 at compile time
- timestamp: merge darwin and freebsd into single notlinux file
- ntp/responder/server: merge darwin and freebsd into single notlinux file
- phc/unix: move CLOCK_MONOTONIC_RAW to linux.go (unavailable on FreeBSD)
- ntpcheck: use phc/unix for CLOCK_MONOTONIC_RAW in track command
- ping_test: remove unnecessary !race build tag
