+++ phc/unix/notlinux.go
@@ -34,9 +34,11 @@ type Timex struct {
 // Linux-only constants that need stub values on Darwin
 //
 //nolint:revive
+//nolint:revive
 const (
-	TIME_OK        = 0
-	CLOCK_BOOTTIME = 7
+	TIME_OK             = 0
+	CLOCK_BOOTTIME      = 7
+	CLOCK_MONOTONIC_RAW = 4
 )
 
 // FdToClockID is not supported on non-linux.
