+++ timestamp/timestamp_notlinux.go
@@ -1,3 +1,5 @@
+//go:build !linux
+
 /*
 Copyright (c) Facebook, Inc. and its affiliates.
 
@@ -18,6 +20,7 @@ package timestamp
 
 import (
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"net"
 	"time"
@@ -26,10 +29,9 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-// unix.Cmsghdr size differs depending on platform
-var socketControlMessageHeaderOffset = binary.Size(unix.Cmsghdr{})
-
 var timestamping = unix.SO_TIMESTAMP
+
+var ErrNoTimestamp = errors.New("no timestamp available")
 
 // Here we have basic HW and SW timestamping support
 
@@ -62,8 +64,8 @@ func socketControlMessageTimestamp(b []byte, _ int) (time.Time, error) {
 	return scmDataToTime(b[unix.CmsgSpace(0):])
 }
 
-// socketControlMessageDrops is not implemented for freebsd
-func socketControlMessageDrops(b []byte, _ int) uint32 {
+// socketControlMessageDrops is not implemented for darwin
+func socketControlMessageDrops(_ []byte, _ int) uint32 {
 	return 0
 }
 
@@ -78,15 +80,45 @@ func EnableTimestamps(ts Timestamp, connFd int, _ *net.Interface) error {
 	switch ts {
 	case SW:
 		if err := EnableSWTimestampsRx(connFd); err != nil {
-			return fmt.Errorf("Cannot enable software timestamps: %w", err)
+			return fmt.Errorf("cannot enable software timestamps: %w", err)
 		}
 	default:
-		return fmt.Errorf("Unrecognized timestamp type: %s", ts)
+		return fmt.Errorf("unrecognized timestamp type: %s", ts)
 	}
 	return nil
 }
 
-// EnableRXQueueOverflow is not implemented for freebsd
-func EnableRXQueueOverflow(connFd int) error {
-	return err.Errorf("RX queue overflow is not supported on freebsd")
+// EnableRXQueueOverflow is not implemented for darwin
+func EnableRXQueueOverflow(_ int) error {
+	return fmt.Errorf("RX queue overflow is not supported on this platform")
+}
+
+// ReadTXtimestamp is not supported on this platform.
+func ReadTXtimestamp(_ int) (time.Time, int, error) {
+	return time.Time{}, 0, errors.New("TX timestamping is not supported on this platform")
+}
+
+// ReadTXtimestampBuf is not supported on this platform.
+func ReadTXtimestampBuf(_ int, _, _ []byte) (time.Time, int, error) {
+	return time.Time{}, 0, errors.New("TX timestamping is unsupported on this platform")
+}
+
+// SeqIDSocketControlMessage is not supported on this platform.
+func SeqIDSocketControlMessage(_ uint32, _ []byte) {
+	// no-op on this platform
+}
+
+// ReadTimeStampSeqIDBuf is not supported on this platform.
+func ReadTimeStampSeqIDBuf(_ int, _ []byte, _ uint32) (time.Time, int, error) {
+	return time.Time{}, 0, errors.New("timestamp sequence ID control messages are unsupported on this platform")
+}
+
+// EnableHWTimestamps is not supported on this platform.
+func EnableHWTimestamps(_ int, _ *net.Interface) error {
+	return errors.New("hardware timestamping is not supported on this platform")
+}
+
+// EnableHWTimestampsRx is not supported on this platform.
+func EnableHWTimestampsRx(_ int, _ *net.Interface) error {
+	return errors.New("hardware RX timestamping is not supported on this platform")
 }
