+++ timestamp/timestamp_notlinux_test.go
@@ -1,3 +1,5 @@
+//go:build !linux
+
 /*
 Copyright (c) Facebook, Inc. and its affiliates.
 
@@ -79,7 +81,7 @@ func TestEnableTimestamps(t *testing.T) {
 
 	// HARDWARE
 	err = EnableTimestamps(HW, connFd, &net.Interface{Name: "lo", Index: 1})
-	require.Equal(t, fmt.Errorf("Unrecognized timestamp type: %s", HW), err)
+	require.Equal(t, fmt.Errorf("unrecognized timestamp type: %s", HW), err)
 }
 
 func TestReadPacketWithRXTimestamp(t *testing.T) {
